Avoid the need to specify custom public path (hack).
This copies the resources to the root projects directory, so that "./public" is usually always the correct path. However, this makes the files available twice, once as "/diskuto.js" and once as "/diskuto/diskuto.js".

It feels like the best final option may be to use string imports to directly include the files into the compiled executable instead of jumping through hoops to get the resource paths right.